% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapPCA.R
\name{bootstrapPCA}
\alias{bootstrapPCA}
\title{Estimate flow across Tamiami Trail}
\usage{
bootstrapPCA(data, probTrain = 0.8, iter = 1000, plot = TRUE,
  columnsUsed)
}
\arguments{
\item{data}{input data (will be subsampled to create training and out-of-sample data)}

\item{probTrain}{proportion of sample data to include in training dataset}

\item{iter}{number of iterations}

\item{plot}{logical; whether relationship between preicted and observed flow be displayed on a plot}

\item{columnsUsed}{numeric range specifying columns in "data" to be used for PCA}
}
\value{
a vector of flow estimates
}
\description{
Applies principal component analysis to new data to generate predicted flow into Everglades National Park
}
