% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFlow.R
\name{estimateFlow}
\alias{estimateFlow}
\title{Estimate flow across Tamiami Trail}
\usage{
estimateFlow(PCA_output, periodOfRecord, newData)
}
\arguments{
\item{PCA_output}{output of principal component analysis}

\item{periodOfRecord}{period of record data used to run PCA}

\item{newData}{a new dataset with identical variable names as the data used for the PCA. Generate this object with the mergeData() function.}
}
\value{
a vector of flow estimates
}
\description{
Applies principal component analysis to new data to generate predicted flow into Everglades National Park
}
