% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDateOffset.R
\name{getDateOffset}
\alias{getDateOffset}
\title{getDateOffser}
\usage{
getDateOffset(dataset, dateColumn = "date", day = "Friday")
}
\arguments{
\item{dataset}{input dataframe}

\item{dateColumn}{name of column with datestamp (Date or POSIX data type)}

\item{day}{character element specifying the first day of each week (default = "Friday", purely to match SFWMD's arbitrary choice)}
}
\value{
a vector of values
}
\description{
Identify the number of days needed to convert between a week starting on "Monday" and the desired start day.
}
\examples{
  dateDat        <- data.frame(date = seq.Date(from = Sys.Date() - 100, to = Sys.Date(), by = 1))
  dateDat$dow    <- weekdays(dateDat$date)
  dateDat$week.R <- format(as.Date(dateDat$date), "\%Y-\%W")
  head(dateDat, 10)
  
  dateOffset        <- getDateOffset(dateDat, day = "Friday")
  dateDat$week.new  <- format(as.Date(dateDat$date) - dateOffset, "\%Y-\%W")
  head(dateDat, 10)
  

}
