% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Merge COP data}
\usage{
mergeData(rainfall, PET, stage, flow)
}
\arguments{
\item{rainfall}{file address of a tab-delimited .txt file containing daily rainfall data}

\item{PET}{file address of a tab-delimited .txt file containing daily PET data}

\item{stage}{file address of a tab-delimited .txt file containing daily stage data}

\item{flow}{file address of a tab-delimited .txt file containing daily flow data}
}
\value{
a dataframe with all input variables
}
\description{
Merges and processes daily rainfall, PET, stage, and flow data into a single dataset aggregated to weekly time steps
}
