% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca1.R, R/por.R
\docType{data}
\name{por}
\alias{por}
\alias{pca1}
\title{Data: principal component analysis model for period of record}
\format{results of a call to FactoMineR::PCA():
\describe{
\item{eig}{eigenvalues for the PCA}
\item{var}{variable names and }
\item{ind}{lagged flow data}
\item{svd}{the response variable: the sum of observed flows across Tamiami Trail}
\item{call}{}
}}
\usage{
pca1

por
}
\description{
A list with 5 elements.

An xts object with 2139 rows and 56 columns. Data have been aggregated from daily to weekly time intervals.
}
\examples{
summary(pca1)

\dontrun{
### code used to generate object 
colsToUse <- 2:55
pca.dat   <- por

pca1 <- FactoMineR::PCA(pca.dat[, colsToUse],  ncp = 4, scale.unit = TRUE)
# save("pca1", file =  paste0(here(), "/data/pca1.RData"))

}

names(por)

\dontrun{
### code used to generate object:
rainDat  <- paste0(here(), "/inst/extdata/data_Daily_Rain.txt")
PETDat   <- paste0(here(), "/inst/extdata/data_daily_PET.txt")
stageDat <- paste0(here(), "/inst/extdata/data_ALT_output_stage.txt")
flowDat  <- paste0(here(), "/inst/extdata/data_ALTO_output_Flows.txt")
por <- mergeData(rainfall = rainDat, PET = PETDat, stage = stageDat, flow = flowDat)
}

}
\keyword{PET}
\keyword{flow}
\keyword{rainfall}
\keyword{stage}
